+++ srcs/vectors_funcs/vector_normalise.c
@@ -1,21 +1,9 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   vector_normalise.c                                 :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: ehillman <ehillman@student.42.fr>          +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/03/20 19:37:49 by ehillman          #+#    #+#             */
-/*   Updated: 2021/03/20 19:38:32 by ehillman         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../includes/minirt.h"
+#include "../includes/MiniRT.h"
 
 t_vector		vector_normalise(t_vector v)
 {
-	float		len;
-	float		len_coeff;
+	double		len;
+	double		len_coeff;
 	t_vector	n;
 
 	n = new_vector(0, 0, 0);
